refactor(sempai-yaml): extract ProjectDependsOnPayload to its own module
Moved the ProjectDependsOnPayload struct and its implementations from model.rs into a new dedicated module project_depends_on.rs. Updated model.rs to re-export ProjectDependsOnPayload and reference the new module. This improves code organization and separation of concerns in the sempai-yaml crate.

Also updated related tests to use the new module properly and adapted execplan documentation to reflect the new type usage.
